requirements.txt
aiogram==3.23.0
aiosqlite==0.20.0
python-dotenv==1.0.1
PyYAML==6.0.2
beautifulsoup4==4.12.3
Pillow==10.4.0
Jinja2==3.1.4
